Move sentry to lib/services
